requirements.txt
# Core dependencies for AI Agent
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
